requirements.txt
instagrapi
spotipy
python-dotenv
Pillow>=8.1.1
pytz